PUT Person fix

+++ src/main/java/home/holymiko/InvestmentScraperApp/Server/Type/DTO/advanced/PersonAccountDTO.java
@@ -9,7 +9,7 @@ public class PersonAccountDTO extends PersonDTO {
 
     private final AccountDTO account;
 
-    public PersonAccountDTO(Long id, String firstName, String middleName, String lastName, String email, int phone, AccountDTO account) {
+    public PersonAccountDTO(Long id, String firstName, String middleName, String lastName, String email, Long phone, AccountDTO account) {
         super(id, firstName, middleName, lastName, email, phone);
         this.account = account;
     }
